📝 (pgvector.py): add support for parsing and transforming connection string in PGVectorStoreComponent to improve security and maintainability 📝 (connection_string_parser.py): create utility function to transform connection string by encoding password to improve security 📝 (test_connection_string_parser.py): add unit tests for transform_connection_string function to ensure correct transformation of connection string
