+++ src/backend/base/langflow/components/vectorstores/pgvector.py
@@ -6,6 +6,7 @@ from langflow.base.vectorstores.model import LCVectorStoreComponent
 from langflow.helpers.data import docs_to_data
 from langflow.io import HandleInput, IntInput, StrInput, SecretStrInput, DataInput, MultilineInput
 from langflow.schema import Data
+from langflow.utils.connection_string_parser import transform_connection_string
 
 
 class PGVectorStoreComponent(LCVectorStoreComponent):
@@ -46,18 +47,20 @@ class PGVectorStoreComponent(LCVectorStoreComponent):
             else:
                 documents.append(_input)
 
+        connection_string_parsed = transform_connection_string(self.pg_server_url)
+
         if documents:
             pgvector = PGVector.from_documents(
                 embedding=self.embedding,
                 documents=documents,
                 collection_name=self.collection_name,
-                connection_string=self.pg_server_url,
+                connection_string=connection_string_parsed,
             )
         else:
             pgvector = PGVector.from_existing_index(
                 embedding=self.embedding,
                 collection_name=self.collection_name,
-                connection_string=self.pg_server_url,
+                connection_string=connection_string_parsed,
             )
 
         return pgvector
